jit: partial x86 support

+++ NULLC/CodeGen_X86.cpp
@@ -741,7 +741,7 @@ void EMIT_OP_REG_NUM(CodeGenGenericContext &ctx, x86Command op, x86Reg reg1, uns
 	ctx.x86Op++;
 }
 
-void EMIT_OP_REG_NUM64(CodeGenGenericContext &ctx, x86Command op, x86Reg reg1, uintptr_t num)
+void EMIT_OP_REG_NUM64(CodeGenGenericContext &ctx, x86Command op, x86Reg reg1, unsigned long long num)
 {
 #ifdef NULLC_OPTIMIZE_X86
 	switch(op)
